+++ jest.config.js
@@ -1,6 +1,6 @@
+/** @type {import('jest').Config} */
 module.exports = {
-  extensionsToTreatAsEsm: ['.ts'], // .tsファイルをECMAScriptモジュールとして扱う
-  transform: {
-    '^.+\\.ts$': 'ts-jest', // TypeScriptファイルを処理するためにts-jestトランスフォーマを使用
-  },
+    preset: 'ts-jest',
+    testEnvironment: 'node',
+    roots: ['<rootDir>/src'],
 };